+++ guillotina/__init__.py
@@ -36,9 +36,10 @@ app_settings = {
         "guillotina.auth.validators.SaltedHashPasswordValidator",
         "guillotina.auth.validators.JWTValidator"
     ],
-    "default_permission": 'zope.Public',
+    "default_permission": 'guillotina.AccessContent',
     "available_addons": {},
     "api_definition": {},
+    "cors_"
     "cors": {
         "allow_origin": ["http://localhost:8080"],
         "allow_methods": ["GET", "POST", "DELETE", "HEAD", "PATCH", "OPTIONS"],